Added write function for image wrapper. Added 'output' stream-like function to PropertyMap. Fixed some warnings.
git-svn-id: svn+ssh://robotics.gwu.edu/home/rpg/repos/RPG/core/RPG@321 e56edf37-dbb1-46d4-901c-b18800ef280d

+++ Devices/Camera/Drivers/HDMI/HDMIDriver.cpp
@@ -34,7 +34,7 @@ bool HDMIDriver::Capture( std::vector<rpg::ImageWrapper>& vImages )
     }
 
     // now copy the images from the delegate
-    for (int ii = 0; ii < m_nNumImages; ii++) {
+    for ( size_t ii = 0; ii < m_nNumImages; ii++) {
         vImages[ii].Image = cv::Mat(m_nImageHeight,m_nImageWidth, CV_8UC3, buffer);
         //advance the pointer forward
         buffer += m_nImageHeight*m_nImageWidth*4;
